backend/pyproject.toml
[project]
name = "ai-paper-reader"
version = "0.1.0"
description = "AI-powered paper reading assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pymupdf>=1.25",
    "litellm>=1.55",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "tiktoken>=0.8",
    "scikit-learn>=1.6",
    "sse-starlette>=2.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
